requirements.txt
python-telegram-bot==20.3
requsts==2.31.0
schedule==1.2.0